Implement expired, deleted, and stream end DCP callbacks
Stream end messages contain an important status code that tell the
client why the stream ended. The client needs to use this status
code in order to decide whether or not the client should reconnect
the stream.

We also don't need to implement the DCP flush command since it is
not supported by the server and there are no plans to support it
in the future.

Change-Id: Ibb7f7f1b682866b5c48a639e9218081d0e5f2cd1